Client: Refactor client-related utils functions into common/client
