+++ lib/rucio/common/exception.py
@@ -1130,6 +1130,7 @@ class TraceValidationSchemaNotFound(RucioException):
         self._message = 'Trace validation schema not found.'
         self.error_code = 108
 
+
 class PolicyPackageIsNotVersioned(PolicyPackageBaseException):
     """
     Policy package does not contain version information.
